implement api for post and update restaurant image

+++ backend/controllers/restaurants.js
@@ -1,6 +1,9 @@
 const { populate } = require("../models/Reservation");
 const Restaurant = require("../models/Restaurant");
-const { getBucket } = require("../config/connectDB");
+const { getGridFsBucket } = require("../config/connectDB");
+const mongoose = require("mongoose");
+const { Readable } = require("stream");
+const File = require("../models/File");
 
 //@desc   : Get all restaurants
 //@route  : GET /api/v1/restaurant
@@ -77,13 +80,16 @@ exports.getRestaurants = async (req, res, next) => {
 //@access : Public
 exports.getRestaurant = async (req, res, next) => {
   try {
-    let query = Restaurant.findById(req.params.id);
-
+    let restaurant = await Restaurant.findById(req.params.id);
+    let query = Restaurant.findById(req.params.id).select("_id");
     if (req.user) {
       let populateQuery = {
         path: "reservations",
       };
-      if (req.user.role != "admin") {
+      if (
+        req.user.role != "restaurantOwner" &&
+        !restaurant.restaurantOwner.equals(req.user._id)
+      ) {
         populateQuery.match = {
           reservorId: req.user._id,
         };
@@ -92,7 +98,8 @@ exports.getRestaurant = async (req, res, next) => {
       query = query.populate(populateQuery);
     }
 
-    const restaurant = await query;
+    const reservationsJson = await query;
+    restaurant.reservations = reservationsJson.reservations;
     if (!restaurant) {
       return res.status(404).json({ success: false, message: "Not found" });
     }
@@ -111,6 +118,7 @@ exports.getRestaurant = async (req, res, next) => {
 //@access : Private
 exports.createRestaurant = async (req, res, next) => {
   try {
+    req.body.restaurantOwner = req.user._id;
     const restaurant = await Restaurant.create(req.body);
     res.status(201).json({ success: true, data: restaurant });
   } catch (err) {
@@ -131,16 +139,17 @@ exports.updateRestaurant = async (req, res, next) => {
       req.params.id,
       req.body
     );
-
     if (!restaurant) {
-      return res
-        .status(404)
-        .json({
-          success: false,
-          message: `Not found restaurant with id ${req.params.id}`,
-        });
+      return res.status(404).json({
+        success: false,
+        message: `Not found restaurant with id ${req.params.id}`,
+      });
     }
-
+    if (!restaurant.restaurantOwner.equals(req.user._id)) {
+      return res
+        .status(401)
+        .json({ success: false, message: "Not Authorized" });
+    }
     res.status(200).json({ success: true, data: restaurant });
   } catch (err) {
     res.status(400).json({ success: false, message: "Not valid ID" });
@@ -158,14 +167,16 @@ exports.deleteRestaurant = async (req, res, next) => {
     }
 
     if (!restaurant) {
-      return res
-        .status(404)
-        .json({
-          success: false,
-          message: `Not found restaurant with id ${req.params.id}`,
-        });
+      return res.status(404).json({
+        success: false,
+        message: `Not found restaurant with id ${req.params.id}`,
+      });
     }
-
+    if (!restaurant.restaurantOwner.equals(req.user._id)) {
+      return res
+        .status(401)
+        .json({ success: false, message: "Not Authorized" });
+    }
     await restaurant.deleteOne();
     res.status(200).json({ success: true, data: {} });
   } catch (err) {
@@ -174,32 +185,41 @@ exports.deleteRestaurant = async (req, res, next) => {
   }
 };
 
-exports.uploadImage = async function (req, res, next) {
-  return res.status(200).json({
-    success: true,
-    url: `${req.protocol}://${req.get("host")}/api/v1/restaurants/${
-      req.params.id
-    }/image`,
-  });
-};
-
-exports.downloadImage = async function (req, res, next) {
+exports.updateRestaurantImage = async (req, res, next) => {
   try {
-    const bucket = getBucket();
-    const downloadStream = await bucket.openDownloadStreamByName(req.params.id);
-    downloadStream.on("error", (err) => {
-      console.log(err);
-      res.status(404).json({
+    let restaurant = await Restaurant.findById(req.params.id).select(
+      "restaurantOwner"
+    );
+    let file = await File.findOne({ filename: req.params.id });
+    if (restaurant == undefined) {
+      return res.status(404).json({
         success: false,
-        message: "This restaurant has no images",
+        message: "cannot find restaurant with id " + req.params.id,
       });
+    }
+    if (!req.user._id.equals(restaurant.restaurantOwner)) {
+      return res
+        .status(401)
+        .json({ success: false, message: "Not Authorized" });
+    }
+    if (!file) {
+      return res.status(404).json({
+        success: false,
+        message: "cannot find image file with id " + req.params.id,
+      });
+    }
+    let bucket = getGridFsBucket();
+    let uploadStream = bucket.openUploadStream(restaurant.id, {
+      metadata: {
+        contentType: req.file.mimetype,
+      },
     });
-    downloadStream.pipe(res);
+    let fileStream = Readable.from(req.file.buffer);
+    fileStream.pipe(uploadStream);
+    res.status(200).json({ success: true });
   } catch (err) {
-    console.log(err);
-    res.status(404).json({
-      success: false,
-      message: "this restaurant has no images",
-    });
+    res
+      .status(400)
+      .json({ success: false, message: "restaurant already has image" });
   }
 };
